refactor : user package 구조 수정
