+++ src/main/java/com/example/book_your_seat/user/service/command/AddressCommandService.java
@@ -2,12 +2,28 @@ package com.example.book_your_seat.user.service.command;
 
 import com.example.book_your_seat.user.controller.dto.AddAddressRequest;
 import com.example.book_your_seat.user.controller.dto.AddressIdResponse;
+import com.example.book_your_seat.user.domain.Address;
 import com.example.book_your_seat.user.domain.User;
+import com.example.book_your_seat.user.repository.AddressRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
-public interface AddressCommandService {
+@Service
+@RequiredArgsConstructor
+public class AddressCommandService {
 
-    AddressIdResponse addAddress(User user, AddAddressRequest addAddressRequest);
+    private final AddressRepository addressRepository;
 
-    AddressIdResponse deleteAddress(Long addressId);
+    public AddressIdResponse addAddress(User user, AddAddressRequest addAddressRequest) {
+
+        Address address = new Address(addAddressRequest.postcode(), addAddressRequest.detail(), user);
+        Address savedAddress = addressRepository.save(address);
+        return new AddressIdResponse(savedAddress.getId());
+    }
+
+    public AddressIdResponse deleteAddress(Long addressId) {
+        addressRepository.deleteById(addressId);
+        return new AddressIdResponse(addressId);
+    }
 
 }
