+++ src/main/java/com/example/book_your_seat/user/service/query/UserQueryService.java
@@ -2,18 +2,45 @@ package com.example.book_your_seat.user.service.query;
 
 import com.example.book_your_seat.user.controller.dto.AddressResponse;
 import com.example.book_your_seat.user.domain.User;
+import com.example.book_your_seat.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface UserQueryService {
+import static com.example.book_your_seat.coupon.CouponConst.USER_NOT_FOUND;
+import static com.example.book_your_seat.user.UserConst.ALREADY_JOIN_EMAIL;
 
-    User getUserByUserId(Long userId);
+@Service
+@RequiredArgsConstructor
+public class UserQueryService {
 
-    User getUserWithUserCoupons(Long userId);
+    private final UserRepository userRepository;
 
-    List<AddressResponse> getUserAddressList(Long userId);
+    public User getUserByUserId(Long userId) {
+        return userRepository.findById(userId)
+                .orElseThrow(() -> new IllegalArgumentException(USER_NOT_FOUND));
+    }
 
-    void checkEmail(String email);
+    public User getUserWithUserCoupons(Long userId) {
+        return userRepository.findByIdWithUserCoupons(userId)
+                .orElseThrow(() -> new IllegalArgumentException(USER_NOT_FOUND));
+    }
 
+    public List<AddressResponse> getUserAddressList(Long userId) {
+        User user = getUserByUserId(userId);
+        return user.getAddressList().stream()
+                .map(address -> new AddressResponse(
+                        address.getPostcode(),
+                        address.getDetail()
+                )).toList();
+    }
+
+    public void checkEmail(String email) {
+        if (userRepository.existsByEmail(email)) {
+            throw new IllegalArgumentException(ALREADY_JOIN_EMAIL);
+        }
+    }
 
 }
